+++ src/Altinn.Apps/AppFrontend/react/altinn-app-frontend/__mocks__/mocks.ts
@@ -1,5 +1,5 @@
 export { getFormLayoutStateMock } from './formLayoutStateMock';
 export { getInitialStateMock } from './initialStateMock';
 export { getFormDataStateMock } from './formDataStateMock';
-export { getLanguageFromCode as getLanguageMock } from '../src/shared/resources/language/languages';
+export { getLanguageFromCode as getLanguageMock } from '../../shared/src/language';
 export { applicationMetadataMock } from './applicationMetadataMock';
