+++ src/Altinn.Apps/AppFrontend/react/shared/src/language/index.d.ts
@@ -1,0 +1,1 @@
+export { getLanguageFromCode } from './languages';
