+++ src/Altinn.Apps/AppFrontend/react/shared/src/language/languages.ts
