+++ src/Altinn.Apps/AppFrontend/react/shared/src/language/texts/nb.ts
